feat: Implement modular structure for PR AI CLI by separating functionality into distinct modules for AI content generation, Git operations, GitHub interactions, and prerequisite checks. Enhance user experience with improved prompts and error handling.
