+++ src/theme.ts
@@ -1,0 +1,13 @@
+import chalk from "chalk";
+
+/**
+ * Chalk theme for consistent console output colors.
+ */
+export const theme = {
+	primary: chalk.blue,
+	success: chalk.green,
+	warning: chalk.yellow,
+	error: chalk.red,
+	info: chalk.cyan,
+	dim: chalk.dim,
+};
