+++ src/ui.ts
@@ -1,0 +1,68 @@
+import inquirer from "inquirer";
+import { theme } from "./theme";
+
+/**
+ * Presents the generated PR content to the user for review and editing.
+ * Allows confirming, editing title/body, or cancelling.
+ * @param {string} initialTitle - The initial title generated by AI.
+ * @param {string} initialBody - The initial body generated by AI.
+ * @returns {Promise<{ title: string; body: string } | null>} - The final title/body object, or null if cancelled.
+ */
+export async function reviewAndConfirmPr(
+	initialTitle: string,
+	initialBody: string,
+): Promise<{ title: string; body: string } | null> {
+	let currentTitle = initialTitle;
+	let currentBody = initialBody;
+
+	while (true) {
+		console.log(`\n${theme.primary("🤖 Generated PR Content (Preview):")}`);
+		console.log(`${theme.info("Title:")} ${currentTitle}`);
+		console.log(`${theme.info("Body:")}\n${theme.dim(currentBody)}`);
+
+		const { action } = await inquirer.prompt([
+			{
+				type: "list",
+				name: "action",
+				message: "Review the generated content:",
+				choices: [
+					{ name: "✅ Confirm and Create PR", value: "confirm" },
+					{ name: "✏️ Edit Title", value: "edit_title" },
+					{ name: "📝 Edit Body (in $EDITOR)", value: "edit_body" },
+					{ name: "❌ Cancel", value: "cancel" },
+				],
+			},
+		]);
+
+		switch (action) {
+			case "confirm":
+				return { title: currentTitle, body: currentBody };
+			case "edit_title": {
+				const { newTitle } = await inquirer.prompt([
+					{
+						type: "input",
+						name: "newTitle",
+						message: "Enter the new PR title:",
+						default: currentTitle,
+					},
+				]);
+				currentTitle = newTitle;
+				break;
+			}
+			case "edit_body": {
+				const answers = await inquirer.prompt({
+					type: "editor",
+					name: "newBody",
+					message: "Edit the PR body (save and close editor to confirm):",
+					default: currentBody,
+					waitForUseInput: true, // Correct property name
+				});
+				currentBody = answers.newBody;
+				break;
+			}
+			case "cancel":
+				console.log(theme.warning("PR creation cancelled by user."));
+				return null;
+		}
+	}
+}
